intel: avs: Update topology file for Digital Microphone Array
It was reported that, after recent update DMIC topology can't be loaded
on some kernels. [1] This is caused by update to how kernel handles
multiple channel on volume controls. However in case of DMIC it is not
really needed. Rebuild DMIC topology and manually edit conf file, before
converting to binary to replace reference to custom callback:
info "257"
with built-in:
info "1"
and then compile to binary.

The result is topology file, which is almost identical as previously
shared one, but with above change:
 00001820  20 56 6f 6c 75 6d 65 00  00 00 00 00 00 00 00 00  | Volume.........|
 00001830  00 00 00 00 00 00 00 00  00 00 00 00 00 00 00 00  |................|
 00001840  00 00 00 00 00 00 00 00  07 00 00 00 01 01 00 00  |................|
-00001850  01 01 00 00 01 01 00 00  00 00 00 00 00 00 00 00  |................|
+00001850  01 01 00 00 01 00 00 00  00 00 00 00 00 00 00 00  |................|
 00001860  00 00 00 00 00 00 00 00  00 00 00 00 00 00 00 00  |................|
 00001870  00 00 00 00 00 00 00 00  00 00 00 00 00 00 00 00  |................|
 00001880  00 00 00 00 00 00 00 00  00 00 00 00 00 00 00 00  |................|

[1]
Link: https://bugzilla.suse.com/show_bug.cgi?id=1243030
Signed-off-by: Amadeusz Sławiński <amadeuszx.slawinski@linux.intel.com>